[[ iOS ]] Refactor iOS specific files to support embedding.
This patch splits up the iOS source files so that those not required
in an embedded context are not included in the main kernel target.

The build-extension-ios.sh tool now decides how to build the simulator
target (whether as a partially linked object, or as a dylib) depending
on the BUILD_EMBEDDED environment variable.

Additionally, it adds an MCIPhoneEmbedded() predicate which returns
true or false depending on the context of operation - this is used
to customize certain iOS startup methods in the embedded context.